+++ examples/api/websocket_fastapi_demo.py
@@ -2,14 +2,7 @@
 
 from fastapi import FastAPI, WebSocket, WebSocketDisconnect
 from fastapi.responses import HTMLResponse
-
-app = FastAPI()
-
-# Import CORS middleware here to keep optional imports localized
-try:
-    from fastapi.middleware.cors import CORSMiddleware  # type: ignore
-except Exception:
-    CORSMiddleware = None  # type: ignore[assignment]
+from fastapi.middleware.cors import CORSMiddleware  # type: ignore
 
 app = FastAPI(title="WOMBAT WebSocket Demo")
 
@@ -22,14 +15,13 @@ if CORSMiddleware is not None:
             "http://127.0.0.1:5173",
             "http://localhost:5174",
             "http://127.0.0.1:5174",
-            "*"
         ],
         allow_credentials=True,
         allow_methods=["*"],
         allow_headers=["*"],
     )
 
-@app.get("/healthz")
+@app.get("/")
 async def health() -> dict[str, str]:
     return {"status": "ok"}
